avm1: Correct `MovieClipLoader.loadClip`
Handle strings, numbers and DisplayObject targets (not just MovieClips).
To support non-MovieClip targets, turn `clip.as_movie_clip().unwrap()`
to `if let Some(mc) = clip.as_movie_clip()` in `Loader`.

+++ core/src/avm1/activation.rs
@@ -2810,7 +2810,7 @@ impl<'a, 'gc, 'gc_context> Activation<'a, 'gc, 'gc_context> {
             let level: DisplayObject<'_> =
                 MovieClip::new(self.base_clip().movie().unwrap(), self.context.gc_context).into();
 
-            level.set_depth(self.context.gc_context, level_id as i32);
+            level.set_depth(self.context.gc_context, level_id);
             level.set_default_root_name(&mut self.context);
             self.context
                 .stage
